fix: validate contractor inn and phone
